==> SpatialTranscriptomics/Analysis/4_clusterin_analysis_main_parkinsons.Rmd <==
---
title: "covid lung clustering - only human data, and differential expression analysis"
author: "Yuvarani Masarapu"
date: "9/13/2021"
output: 
  html_document:
    code_folding: hide
    toc: TRUE
---

```{r}
library(SeuratObject)
library(Seurat)
library(harmony)
library(sctransform)
library(Matrix)
library(RColorBrewer)
library(ggplot2)
```

### Load filtered seurat object
```{r}
data.object.filter.norm <- readRDS(file = "parkinsons_brain_norm.rds")
```

### Data integration
```{r eval=FALSE}
integration.features <- SelectIntegrationFeatures(object.list = data.object.filter.norm, nfeatures = 5000)

#modifying the maximum size of global variables
options(future.globals.maxSize = 8000 * 1024^8)

data.object.filter.norm <- PrepSCTIntegration(object.list = data.object.filter.norm, anchor.features = integration.features, verbose = FALSE)

set.seed(73)
integration.anchors <- FindIntegrationAnchors(object.list = data.object.filter.norm, normalization.method = "SCT", anchor.features = integration.features, verbose = FALSE, reference = 17) #using sample 17 with highest number of spots as reference

saveRDS(integration.anchors, file = "integration.achors.rds")
#integration.achors <- readRDS("~/Documents/parkinsons_project/analysis-parkinsons/integration.achors.rds")

data.object.filter.norm.integrated = IntegrateData(anchorset = integration.anchors, normalization.method = "SCT", verbose = FALSE)
saveRDS(data.object.filter.norm.integrated, file = "data.integrated_updated.rds")

VariableFeatures(data.object.filter.norm.integrated) <- integration.features

data.object.filter.norm.integrated <- RunPCA(data.object.filter.norm.integrated, assay = "SCT", verbose = FALSE, features = integration.features, npcs = 50, reduction.name = "pca_before_harmony", seed.use = 61)

ElbowPlot(object = data.object.filter.norm.integrated, ndims = 50, reduction = "pca_before_harmony") + ggtitle("Elbow plot to select significant PCs")
```

Adding mouse ID
```{r echo=FALSE}
data.object.filter.norm.integrated$mouse_id <- colnames(data.object.filter.norm.integrated)

data.object.filter.norm.integrated$mouse_id[data.object.filter.norm.integrated$sample_id %in% c("V11D08-305_A1" , "V11D08-305_B1")] <- "534097"
data.object.filter.norm.integrated$mouse_id[data.object.filter.norm.integrated$sample_id %in% c("V11D08-305_C1" , "V11D08-305_D1")] <- "534100"
data.object.filter.norm.integrated$mouse_id[data.object.filter.norm.integrated$sample_id %in% c("V11D08-306_A1" , "V11D08-306_B1")] <- "534100"
data.object.filter.norm.integrated$mouse_id[data.object.filter.norm.integrated$sample_id %in% c("V11D08-306_C1" , "V11D08-306_D1")] <- "534097"
data.object.filter.norm.integrated$mouse_id[data.object.filter.norm.integrated$sample_id %in% c("V11D06-069_A1" , "V11D06-069_B1")] <- "534094"
data.object.filter.norm.integrated$mouse_id[data.object.filter.norm.integrated$sample_id %in% c("V11D06-069_C1" , "V11D06-069_D1")] <- "534107"
data.object.filter.norm.integrated$mouse_id[data.object.filter.norm.integrated$sample_id %in% c("V11D06-070_A1" , "V11D06-070_B1")] <- "534107"
data.object.filter.norm.integrated$mouse_id[data.object.filter.norm.integrated$sample_id %in% c("V11D06-070_C1" , "V11D06-070_D1")] <- "534094"

data.object.filter.norm.integrated$mouse_id[data.object.filter.norm.integrated$sample_id %in% c("V11D06-099_A1")] <- "534103"
data.object.filter.norm.integrated$mouse_id[data.object.filter.norm.integrated$sample_id %in% c("V11D06-099_B1")] <- "534095"
data.object.filter.norm.integrated$mouse_id[data.object.filter.norm.integrated$sample_id %in% c("V11D06-099_C1")] <- "534103"
data.object.filter.norm.integrated$mouse_id[data.object.filter.norm.integrated$sample_id %in% c("V11D06-099_D1")] <- "534095"

data.object.filter.norm.integrated$mouse_id[data.object.filter.norm.integrated$sample_id %in% c("V11D06-100_A1")] <- "534095"
data.object.filter.norm.integrated$mouse_id[data.object.filter.norm.integrated$sample_id %in% c("V11D06-100_B1")] <- "534103"
data.object.filter.norm.integrated$mouse_id[data.object.filter.norm.integrated$sample_id %in% c("V11D06-100_C1")] <- "534095"
data.object.filter.norm.integrated$mouse_id[data.object.filter.norm.integrated$sample_id %in% c("V11D06-100_D1")] <- "534103"
```


### Dimensionality reduction and Clustering

 *35 dimensions chosen for further analysis

```{r eval=FALSE}
DefaultAssay(data.object.filter.norm.integrated) <- "SCT"

#modifying the maximum size of global variables
options(future.globals.maxSize = 8000 * 1024^2)

data.object.filter.norm.integrated <- RunUMAP(object = data.object.filter.norm.integrated, dims = 1:35, assay = "SCT", seed.use = 7913, reduction = "pca_before_harmony", reduction.name = "umap_before_harmony") #35 contributing PCs chosen for further analysis

DimPlot(object = data.object.filter.norm.integrated, group.by = "sample_id", reduction = "umap_before_harmony") + ggtitle("Before batch correction (spots grouped by sections)")
#slides well blended
DimPlot(object = data.object.filter.norm.integrated, group.by = "slide_id", reduction = "umap_before_harmony")  + ggtitle("slides")
#sub arrays don't blend well
DimPlot(object = data.object.filter.norm.integrated, group.by = "mouse_id", reduction = "umap_before_harmony") + ggtitle("mouse id")
DimPlot(object = data.object.filter.norm.integrated, group.by = "sample_type", reduction = "umap_before_harmony") + ggtitle("sample type")
#levels blended well as well
DimPlot(object = data.object.filter.norm.integrated, group.by = "sectioning_level", reduction = "umap_before_harmony") + ggtitle("sectioning level")
DimPlot(object = data.object.filter.norm.integrated, group.by = "sub_array", reduction = "umap_before_harmony") + ggtitle("sub array")
```

```{r}
library(RColorBrewer)
n <- 60
qual_col_pals = brewer.pal.info[brewer.pal.info$category == 'qual',]

col_vector = unlist(mapply(brewer.pal, qual_col_pals$maxcolors, rownames(qual_col_pals)))
```


```{r eval=FALSE}
set.seed(31)
data.object.filter.norm.integrated <- RunHarmony(object = data.object.filter.norm.integrated, group.by.vars = c("sample_id"), theta = c(2), assay.use = "SCT", reduction = "pca_before_harmony", dims.use = 1:35, plot_convergence = T, verbose = F, reduction.save = "harmony_sid")

data.object.filter.norm.integrated <- RunUMAP(object = data.object.filter.norm.integrated, assay.use = "SCT", reduction = "harmony_sid", dims = 1:35, seed.use = 5731, reduction.name = "umap_after_harmony")

DimPlot(object = data.object.filter.norm.integrated, group.by = "sample_id", reduction = "umap_after_harmony", cols = col_vector) + ggtitle("Batch corrected data")
DimPlot(object = data.object.filter.norm.integrated, group.by = "slide_id", reduction = "umap_after_harmony", cols = col_vector) + ggtitle("Batch corrected data")
DimPlot(object = data.object.filter.norm.integrated, group.by = "mouse_id", reduction = "umap_after_harmony", cols = col_vector) + ggtitle("Batch corrected data")
DimPlot(object = data.object.filter.norm.integrated, group.by = "sample_type", reduction = "umap_after_harmony", cols = col_vector) + ggtitle("Batch corrected data")
DimPlot(object = data.object.filter.norm.integrated, group.by = "sectioning_level", reduction = "umap_after_harmony", cols = col_vector) + ggtitle("Batch corrected data")
DimPlot(object = data.object.filter.norm.integrated, group.by = "sub_array", reduction = "umap_after_harmony", cols = col_vector) + ggtitle("Batch corrected data")
FeaturePlot(object = data.object.filter.norm.integrated, features = "nCount_SCT", reduction = "umap_after_harmony", cols = c("light pink" ,"pink", "magenta"))
```

```{r eval=FALSE}
saveRDS(data.object.filter.norm.integrated, file = "data.integrated.harm.rds")
```

```{r eval=FALSE}
set.seed(6741)
data.object.filter.norm.integrated <- FindNeighbors(object = data.object.filter.norm.integrated, assay = "SCT", dims = 1:35, k.param = 23, graph.name = "graph_afterHarmony", reduction = "harmony_sid")
```

Resolution 1.5, 26 clusters
```{r eval=FALSE}
data.object.filter.norm.integrated <- FindClusters(object = data.object.filter.norm.integrated, pc.use = 1:35, resolution = 1.5, save.SNN = T, do.sparse = T, graph.name = "graph_afterHarmony", random.seed = 3715, group.singletons = TRUE)

#coldef_res0.6 <-  c("#264653", "#2A9D8F" ,"#A8DADC", "#457B9D", "#1D3557", "#606C38" , "#FB8B24" , "#E36414", "#941B0C", "#A59E8C" , "#66635B" , "#0077B6"  , "#4D194D" , "#A7C957" , "#FFD166" , "#DB3A34", "#E0FBFC", "#63372C")
coldef_res1.5 <-  c("#F18F01", "#878787", "#FCCDE5", "#048BA8", "#2E4057", "#99C24D", "#B95F89", "#5F070C" ,  "#DFC27D", "#DB6C79", "#B191FF", "#157A6E", "#73683B", "#97C8EB","#C51B7D", "#BA9763", "#31081F", "#52D1DC", "#700353", "#F5D3C8", "#725752", "#D8315B", "#6B4E71", "#8CBFCF" , "#C2B2B4", "#EDE5A6")

data.object.filter.norm.integrated$clusters_res1.5 <- data.object.filter.norm.integrated$seurat_clusters
p <- DimPlot(object = data.object.filter.norm.integrated, group.by = "seurat_clusters", reduction = "umap_after_harmony", pt.size = 0.8, label = T, cols = coldef_res1.5, label.size = 12) + ggtitle("Clusters in UMAP space (resolution 1.5)")
ggsave(p, filename = paste(dir, "/clustering_results_parkinsons/clusters_res1.5/clusters_umap_res1.5.pdf", sep = ""), dpi = 300, width = 16, height = 16)

p <- VlnPlot(object = data.object.filter.norm.integrated, features = "nFeature_Spatial", group.by = "seurat_clusters", assay = "SCT", pt.size = 0, cols = coldef_res1.5) + ggtitle("Genes/spot per cluster (resolution 1.5)")
ggsave(p, filename = paste(dir, "/clustering_results_parkinsons/clusters_res1.5/genes_per_cluster_res1.5.pdf", sep = ""), dpi = 300, width = 14, height = 10)

p <- VlnPlot(object = data.object.filter.norm.integrated, features = "nCount_Spatial", group.by = "seurat_clusters", assay = "SCT", pt.size = 0, cols = coldef_res1.5) + ggtitle("UMIs/spot per cluster (resolution 1.5)")
ggsave(p, filename = paste(dir, "/clustering_results_parkinsons/clusters_res1.5/umi_per_cluster_res1.5.pdf", sep = ""), dpi = 300, width = 14, height = 10)

#Total number of spots per cluster
cluster_table_res1.5 <- table(data.object.filter.norm.integrated@active.ident, data.object.filter.norm.integrated@meta.data$orig.ident)
write.table(cluster_table_res1.5, file = paste(dir, "/clustering_results_parkinsons/clusters_res1.5/cluster_stats_res1.5.tsv", sep = ""), row.names = T, col.names = T, sep = ";")

#order of images for samples
#first treated, then controls
images <- c("slice1", "slice1.1", "slice1.2", "slice1.3", "slice1.6", "slice1.7", "slice1.12", "slice1.13", "slice1.18", "slice1.19", "slice1.20", "slice1.21",
            "slice1.4","slice1.5","slice1.8","slice1.9","slice1.10","slice1.11", "slice1.14", "slice1.15", "slice1.16", "slice1.17", "slice1.22", "slice1.23")
sample_type <- c(rep("Treated", 12), rep("Control", 12))
sample_name <- c("V11D06-099_A1","V11D06-099_C1","V11D08-305_C1","V11D08-306_A1","V11D06-069_D1","V11D06-100_D1","V11D06-070_A1","V11D06-070_B1","V11D06-069_C1","V11D06-100_B1","V11D08-305_D1","V11D08-306_B1",
                 "V11D06-100_A1","V11D08-305_A1","V11D06-069_A1","V11D06-070_C1","V11D06-099_B1","V11D06-099_D1","V11D06-069_B1","V11D06-070_D1","V11D06-100_C1","V11D08-305_B1","V11D08-306_C1","V11D08-306_D1")

names(coldef_res1.5) <- levels(data.object.filter.norm.integrated)
dir <- getwd()
for(i in 1:length(images)){
  p <- SpatialDimPlot(object = data.object.filter.norm.integrated, crop = F, pt.size.factor = 1, stroke = 0, images = images[i], label = T, cols = coldef_res1.5) +
    ggtitle(paste(unique(sample_type[i])," ID: ", unique(sample_name[i]),sep = "" )) + 
    theme(plot.title = element_text(size = 40, hjust = 0.5), 
          title = element_text(size = 30), 
          legend.key.size = unit(20, units = "mm"),
          legend.text = element_text(size = 20), 
          legend.direction = "vertical", 
          legend.position = "right", 
          legend.key = element_rect(size = 60), 
          legend.key.height = unit(10, units = "mm"))
  ggsave(p, filename = paste(dir, "/clustering_results_parkinsons/clusters_res1.5/clusters_on_tissue/", sample_name[i], "_clusters_res1.5.pdf", sep = ""), dpi = 300, width = 14, height = 14)
}

data.object.filter.norm.integrated <- PrepSCTFindMarkers(object = data.object.filter.norm.integrated, assay = "SCT", verbose = TRUE)

markers_res1.5 <- FindAllMarkers(object = data.object.filter.norm.integrated, assay = "SCT", min.pct = 0.1, only.pos = FALSE)
write.csv(markers_res1.5, file = paste(dir,"/clustering_results_parkinsons/clusters_res1.5/cluster_markers_res1.5.csv", sep = ""), row.names = TRUE, col.names = TRUE)

library(dplyr)
temp_markers_res1.5 <- split(markers_res1.5, f = markers_res1.5$cluster)
for(i in 1:length(unique(data.object.filter.norm.integrated$clusters_res1.5))){
  temp_markers_res1.5[[i]] <- temp_markers_res1.5[[i]][order(temp_markers_res1.5[[i]]$avg_log2FC, decreasing = TRUE),]
}

library(openxlsx)
write.xlsx(temp_markers_res1.5, file = paste(dir,"/clustering_results_parkinsons/clusters_res1.5/cluster_markers_res1.5.xlsx", sep = ""))

data.object.filter.norm.integrated$clusters_res1.5 <- data.object.filter.norm.integrated$seurat_clusters
```

Resolution 2.0, 32 clusters
```{r}
data.object.filter.norm.integrated <- FindClusters(object = data.object.filter.norm.integrated, pc.use = 1:35, resolution = 2.0, save.SNN = T, do.sparse = T, graph.name = "graph_afterHarmony", random.seed = 43, group.singletons = TRUE)

coldef_res2.0 <-  c("#F18F01", "#878787", "#FCCDE5", "#048BA8", "#2E4057", "#99C24D", "#B95F89", "#5F070C" ,  "#DFC27D", "#DB6C79", "#B191FF", "#157A6E", "#73683B", "#97C8EB","#C51B7D", "#BA9763", "#31081F", "#52D1DC", "#700353", "#F5D3C8", "#725752", "#D8315B", "#6B4E71", "#8CBFCF" , "#C2B2B4", "#EDE5A6", "#52B788", "#EE1B26", "#F2F230", "#91F291", "#386CB0", "#E7298A")
pie(rep(1,length(coldef_res2.0)), col = coldef_res2.0)

p <- DimPlot(object = data.object.filter.norm.integrated, group.by = "seurat_clusters", reduction = "umap_after_harmony", pt.size = 0.8, label = T, cols = coldef_res2.0, label.size = 12) + ggtitle("Clusters in UMAP space (resolution 2.0)")
ggsave(p, filename = paste(dir, "/clustering_results_parkinsons/clusters_res2.0/clusters_umap_res2.0.pdf", sep = ""), dpi = 300, width = 16, height = 16)

p <- VlnPlot(object = data.object.filter.norm.integrated, features = "nFeature_Spatial", group.by = "seurat_clusters", assay = "SCT", pt.size = 0, cols = coldef_res2.0) + ggtitle("Genes/spot per cluster (resolution 2.0)")
ggsave(p, filename = paste(dir, "/clustering_results_parkinsons/clusters_res2.0/genes_per_cluster_res2.0.pdf", sep = ""), dpi = 300, width = 14, height = 10)

p <- VlnPlot(object = data.object.filter.norm.integrated, features = "nCount_Spatial", group.by = "seurat_clusters", assay = "SCT", pt.size = 0, cols = coldef_res2.0) + ggtitle("UMIs/spot per cluster (resolution 2.0)")
ggsave(p, filename = paste(dir, "/clustering_results_parkinsons/clusters_res2.0/umi_per_cluster_res2.0.pdf", sep = ""), dpi = 300, width = 14, height = 10)

#Total number of spots per cluster
cluster_table_res2.0 <- table(data.object.filter.norm.integrated@active.ident, data.object.filter.norm.integrated@meta.data$orig.ident)
write.table(cluster_table_res2.0, file = paste(dir, "/clustering_results_parkinsons/clusters_res2.0/cluster_stats_res2.0.tsv", sep = ""), row.names = T, col.names = T, sep = ";")

#order of images for samples
#first treated, then controls
images <- c("slice1", "slice1.1", "slice1.2", "slice1.3", "slice1.6", "slice1.7", "slice1.12", "slice1.13", "slice1.18", "slice1.19", "slice1.20", "slice1.21",
            "slice1.4","slice1.5","slice1.8","slice1.9","slice1.10","slice1.11", "slice1.14", "slice1.15", "slice1.16", "slice1.17", "slice1.22", "slice1.23")
sample_type <- c(rep("Treated", 12), rep("Control", 12))
sample_name <- c("V11D06-099_A1","V11D06-099_C1","V11D08-305_C1","V11D08-306_A1","V11D06-069_D1","V11D06-100_D1","V11D06-070_A1","V11D06-070_B1","V11D06-069_C1","V11D06-100_B1","V11D08-305_D1","V11D08-306_B1",
                 "V11D06-100_A1","V11D08-305_A1","V11D06-069_A1","V11D06-070_C1","V11D06-099_B1","V11D06-099_D1","V11D06-069_B1","V11D06-070_D1","V11D06-100_C1","V11D08-305_B1","V11D08-306_C1","V11D08-306_D1")

names(coldef_res2.0) <- levels(data.object.filter.norm.integrated)
dir <- getwd()
for(i in 1:length(images)){
  p <- SpatialDimPlot(object = data.object.filter.norm.integrated, crop = F, pt.size.factor = 1, stroke = 0, images = images[i], label = T, cols = coldef_res2.0) +
    ggtitle(paste(unique(sample_type[i])," ID: ", unique(sample_name[i]),sep = "" )) + 
    theme(plot.title = element_text(size = 40, hjust = 0.5), 
          title = element_text(size = 30), 
          legend.key.size = unit(20, units = "mm"),
          legend.text = element_text(size = 20), 
          legend.direction = "vertical", 
          legend.position = "right", 
          legend.key = element_rect(size = 60), 
          legend.key.height = unit(10, units = "mm"))
  ggsave(p, filename = paste(dir, "/clustering_results_parkinsons/clusters_res2.0/clusters_on_tissue/", sample_name[i], "_clusters_res2_0.pdf", sep = ""), dpi = 300, width = 14, height = 14)
}

#data.object.filter.norm.integrated <- PrepSCTFindMarkers(object = data.object.filter.norm.integrated, assay = "SCT", verbose = TRUE)
markers_res2.0 <- FindAllMarkers(object = data.object.filter.norm.integrated, assay = "SCT", min.pct = 0.1, only.pos = FALSE, random.seed = 17)
write.csv(markers_res2.0, file = paste(dir,"/clustering_results_parkinsons/clusters_res2.0/cluster_markers_res2.0.csv", sep = ""), row.names = TRUE, col.names = TRUE)
saveRDS(markers_res2.0, file = paste(dir,"/clustering_results_parkinsons/clusters_res2.0/markers_res2.0.rds", sep = ""))


library(dplyr)
temp_markers_res2.0 <- split(markers_res2.0, f = markers_res2.0$cluster)
for(i in 1:length(unique(data.object.filter.norm.integrated$clusters_res2.0))){
  temp_markers_res2.0[[i]] <- temp_markers_res2.0[[i]][order(temp_markers_res2.0[[i]]$avg_log2FC, decreasing = TRUE),]
}
library(openxlsx)
write.xlsx(temp_markers_res2.0, file = paste(dir,"/clustering_results_parkinsons/clusters_res2.0/cluster_markers_res2.0.xlsx", sep = ""))

#save the clusters
data.object.filter.norm.integrated$clusters_res2.0 <- data.object.filter.norm.integrated$seurat_clusters

### DE analysis for object with 32 clusters, resolution 2.0
# DE genes per cluster, treated vs control
data.object.filter.norm.integrated <- PrepSCTFindMarkers(object = data.object.filter.norm.integrated, assay = "SCT", verbose = TRUE)

clusters.list <- SplitObject(object = data.object.filter.norm.integrated, split.by = "ident")
saveRDS(clusters.list, file = paste(dir,"/clustering_results_parkinsons/clusters_res2.0/object_split_by_cluster.rds", sep = ""))

#Tried reinstalling sctransform develop to address prepsctfindmarkers() issue
#remotes::install_github("ChristophH/sctransform@develop")

#Ran PrepSCTFindMarkers() already to recorrect UMIs in the main dataset and SCT assay already replaced by recorrect counts, data slor replaced by log1p of recorrected counts. Hence not doing running that again here.
#Also writing recorrect_umi=FALSE when invoke FindMarkers() and avoid throwing errors.
de.genes.treated <- lapply(clusters.list, function(cl_obj){
  #cl_obj <- PrepSCTFindMarkers(object = cl_obj, assay = "SCT", verbose = TRUE)
  temp <- FindMarkers(object = cl_obj, assay = "SCT", ident.1 = "Treated", ident.2 = "Control", group.by = "sample_type", min.pct = 0.1, min.cells.group = 3, random.seed = 59, recorrect_umi=FALSE)
})

for(i in 1:length(de.genes.treated)){
  de.genes.treated[[i]]$gene <- rownames(de.genes.treated[[i]])
  de.genes.treated[[i]] <- de.genes.treated[[i]][order(de.genes.treated[[i]]$avg_log2FC, decreasing = TRUE),]
}
#names(de.genes.treated) to get the new order starting clusters 0 to 31
new_order_de <- c(21,1,8,15,14,29,27,20,12,17,9,25,5,10,2,31,4,19,6,13,22,23,18,30,3,16,28,24,32,11,26,7)
de.genes.treated1 <- de.genes.treated[new_order_de]

library(openxlsx)
write.xlsx(de.genes.treated1, file = paste(dir,"/clustering_results_parkinsons/clusters_res2.0/DE_genes_control_vs_treated_per_cluster.xlsx", sep = ""))
```


```{r}
if (!require("BiocManager", quietly = TRUE))
    install.packages("BiocManager")

BiocManager::install("MAST")

markers_res2.0_MAST <- FindAllMarkers(object = data.object.filter.norm.integrated, assay = "SCT", min.pct = 0.1, only.pos = FALSE, random.seed = 17, test.use = "MAST")
write.csv(markers_res2.0_MAST, file = paste(dir,"/clustering_results_parkinsons/clusters_res2.0/cluster_markers_res2.0_MAST.csv", sep = ""), row.names = TRUE, col.names = TRUE)

saveRDS(markers_res2.0_MAST, file = paste(dir,"/clustering_results_parkinsons/clusters_res2.0/markers_res2.0_MAST.rds", sep = ""))

library(dplyr)
temp_markers_res2.0_MAST <- split(markers_res2.0_MAST, f = markers_res2.0_MAST$cluster)
for(i in 1:length(unique(data.object.filter.norm.integrated$seurat_clusters))){
  temp_markers_res2.0_MAST[[i]] <- temp_markers_res2.0_MAST[[i]][order(temp_markers_res2.0_MAST[[i]]$avg_log2FC, decreasing = TRUE),]
}
library(openxlsx)
write.xlsx(temp_markers_res2.0_MAST, file = paste(dir,"/clustering_results_parkinsons/clusters_res2.0/cluster_markers_res2.0_MAST.xlsx", sep = ""))


de.genes.treated <- lapply(clusters.list, function(cl_obj){
  #cl_obj <- PrepSCTFindMarkers(object = cl_obj, assay = "SCT", verbose = TRUE)
  temp <- FindMarkers(object = cl_obj, assay = "SCT", ident.1 = "Treated", ident.2 = "Control", group.by = "sample_type", min.pct = 0.1, min.cells.group = 3, random.seed = 59, recorrect_umi=FALSE, test.use = "MAST")
})

for(i in 1:length(de.genes.treated)){
  de.genes.treated[[i]]$gene <- rownames(de.genes.treated[[i]])
  de.genes.treated[[i]] <- de.genes.treated[[i]][order(de.genes.treated[[i]]$avg_log2FC, decreasing = TRUE),]
}
#names(de.genes.treated) to get the new order starting clusters 0 to 31
new_order_de <- c(21,1,8,15,14,29,27,20,12,17,9,25,5,10,2,31,4,19,6,13,22,23,18,30,3,16,28,24,32,11,26,7)
de.genes.treated1 <- de.genes.treated[new_order_de]

library(openxlsx)
write.xlsx(de.genes.treated1, file = paste(dir,"/clustering_results_parkinsons/clusters_res2.0/DE_genes_control_vs_treated_per_cluster-MAST.xlsx", sep = ""))
```

Compare MAST markers with wilcox cluster markers
```{r}
for(i in 1:32){
  print(unique(rownames(temp_markers_res2.0[[i]]) %in% rownames(temp_markers_res2.0_MAST[[i]])))
}
```

Resolution 2.3, 35 clusters
```{r}
data.object.filter.norm.integrated <- FindClusters(object = data.object.filter.norm.integrated, pc.use = 1:35, resolution = 2.3, save.SNN = T, do.sparse = T, graph.name = "graph_afterHarmony", random.seed = 43, group.singletons = TRUE)

coldef_res2.3 <-  c("#F18F01", "#878787", "#FCCDE5", "#048BA8", "#2E4057", "#99C24D", "#B95F89", "#5F070C" ,  "#DFC27D", "#DB6C79", "#B191FF", "#157A6E", "#EB4511", "#73683B", "#97C8EB","#C51B7D", "#BA9763", "#31081F", "#52D1DC", "#700353", "#F5D3C8", "#725752", "#D8315B", "#6B4E71", "#8CBFCF" , "#C2B2B4", "#690375" ,"#EDE5A6", "#52B788", "#EE1B26", "#F2F230", "#91F291", "#845A6D" ,"#386CB0", "#E7298A")
pie(rep(1,length(coldef_res2.3)), col = coldef_res2.3)

p <- DimPlot(object = data.object.filter.norm.integrated, group.by = "seurat_clusters", reduction = "umap_after_harmony", pt.size = 0.8, label = T, label.size = 12, cols = coldef_res2.3) + ggtitle("Clusters in UMAP space (resolution 2.3)")
ggsave(p, filename = paste(dir, "/clustering_results_parkinsons/clusters_res2.3/clusters_umap_res2.3.pdf", sep = ""), dpi = 300, width = 16, height = 16)

p <- VlnPlot(object = data.object.filter.norm.integrated, features = "nFeature_Spatial", group.by = "seurat_clusters", assay = "SCT", pt.size = 0, cols = coldef_res2.3) + ggtitle("Genes/spot per cluster (resolution 2.3)")
ggsave(p, filename = paste(dir, "/clustering_results_parkinsons/clusters_res2.3/genes_per_cluster_res2.3.pdf", sep = ""), dpi = 300, width = 14, height = 10)

p <- VlnPlot(object = data.object.filter.norm.integrated, features = "nCount_Spatial", group.by = "seurat_clusters", assay = "SCT", pt.size = 0, cols = coldef_res2.3) + ggtitle("UMIs/spot per cluster (resolution 2.3)")
ggsave(p, filename = paste(dir, "/clustering_results_parkinsons/clusters_res2.3/umi_per_cluster_res2.3.pdf", sep = ""), dpi = 300, width = 14, height = 10)

#Total number of spots per cluster
cluster_table_res2.3 <- table(data.object.filter.norm.integrated@active.ident, data.object.filter.norm.integrated@meta.data$orig.ident)
write.table(cluster_table_res2.3, file = paste(dir, "/clustering_results_parkinsons/clusters_res2.3/cluster_stats_res2.3.tsv", sep = ""), row.names = T, col.names = T, sep = ";")

#order of images for samples
#first treated, then controls
images <- c("slice1", "slice1.1", "slice1.2", "slice1.3", "slice1.6", "slice1.7", "slice1.12", "slice1.13", "slice1.18", "slice1.19", "slice1.20", "slice1.21",
            "slice1.4","slice1.5","slice1.8","slice1.9","slice1.10","slice1.11", "slice1.14", "slice1.15", "slice1.16", "slice1.17", "slice1.22", "slice1.23")
sample_type <- c(rep("Treated", 12), rep("Control", 12))
sample_name <- c("V11D06-099_A1","V11D06-099_C1","V11D08-305_C1","V11D08-306_A1","V11D06-069_D1","V11D06-100_D1","V11D06-070_A1","V11D06-070_B1","V11D06-069_C1","V11D06-100_B1","V11D08-305_D1","V11D08-306_B1",
                 "V11D06-100_A1","V11D08-305_A1","V11D06-069_A1","V11D06-070_C1","V11D06-099_B1","V11D06-099_D1","V11D06-069_B1","V11D06-070_D1","V11D06-100_C1","V11D08-305_B1","V11D08-306_C1","V11D08-306_D1")

names(coldef_res2.3) <- levels(data.object.filter.norm.integrated)
dir <- getwd()
for(i in 1:length(images)){
  p <- SpatialDimPlot(object = data.object.filter.norm.integrated, crop = F, pt.size.factor = 1, stroke = 0, images = images[i], label = T, cols = coldef_res2.3) +
    ggtitle(paste(unique(sample_type[i])," ID: ", unique(sample_name[i]),sep = "" )) + 
    theme(plot.title = element_text(size = 40, hjust = 0.5), 
          title = element_text(size = 30), 
          legend.key.size = unit(20, units = "mm"),
          legend.text = element_text(size = 20), 
          legend.direction = "vertical", 
          legend.position = "right", 
          legend.key = element_rect(size = 60), 
          legend.key.height = unit(10, units = "mm"))
  ggsave(p, filename = paste(dir, "/clustering_results_parkinsons/clusters_res2.3/clusters_on_tissue/", sample_name[i], "_clusters_res2.3.pdf", sep = ""), dpi = 300, width = 14, height = 14)
}

#data.object.filter.norm.integrated <- PrepSCTFindMarkers(object = data.object.filter.norm.integrated, assay = "SCT", verbose = TRUE)
markers_res2.3 <- FindAllMarkers(object = data.object.filter.norm.integrated, assay = "SCT", min.pct = 0.1, only.pos = FALSE, random.seed = 19)
write.csv(markers_res2.3, file = paste(dir,"/clustering_results_parkinsons/clusters_res2.3/cluster_markers_res2.3.csv", sep = ""), row.names = TRUE, col.names = TRUE)

library(dplyr)
temp_markers_res2.3 <- split(markers_res2.3, f = markers_res2.3$cluster)
for(i in 1:length(levels(data.object.filter.norm.integrated))){
  temp_markers_res2.3[[i]] <- temp_markers_res2.3[[i]][order(temp_markers_res2.3[[i]]$avg_log2FC, decreasing = TRUE),]
}

library(openxlsx)
write.xlsx(temp_markers_res2.3, file = paste(dir,"/clustering_results_parkinsons/clusters_res2.3/cluster_markers_res2.3.xlsx", sep = ""))

#save the clusters
data.object.filter.norm.integrated$clusters_res2.3 <- data.object.filter.norm.integrated$seurat_clusters

saveRDS(data.object.filter.norm.integrated, file = "data.integrated.clusters_2022-08-22.rds")
```

```{r}
# "Drd1a" not found
markers <- c("Sema5b", "Kremen1", "Tac1", "Spon1", "Tshz1", "Lypd1", "Oprm1", "Id4", "Sgk1", "Epha4", "Col6a1", "Crym" ,"Wfs1", "Dlk1", "Gpr155", "Cnr1", "Drd2", "Adora2a")

lapply(markers, function(gene){
  p <- FeaturePlot(object = data.object.filter.norm.integrated, features = gene, reduction = "umap_after_harmony", pt.size = 0.8)
  ggsave(p, filename = paste(dir, "/umaps_markers/", gene, ".pdf", sep = ""), dpi = 300, width = 16, height = 16)
})

library(plotly)
library(cowplot)
lapply(markers, function(gene){
  p1 <- FeaturePlot(object = data.object.filter.norm.integrated, features = gene, reduction = "umap_after_harmony", pt.size = 0.8, cells = colnames(data.object.filter.norm.integrated)[data.object.filter.norm.integrated$sample_type %in% "Treated"]) + ggtitle(paste(gene, "in Treated"))
  p2 <- DimPlot(object = data.object.filter.norm.integrated, group.by = "sample_type", reduction = "umap_after_harmony", pt.size = 0.8)
  p3 <- FeaturePlot(object = data.object.filter.norm.integrated, features = gene, reduction = "umap_after_harmony", pt.size = 0.8, cells = colnames(data.object.filter.norm.integrated)[data.object.filter.norm.integrated$sample_type %in% "Control"]) + ggtitle(paste(gene, "in Control"))
  p4 <- DimPlot(object = data.object.filter.norm.integrated, group.by = "sample_type", reduction = "umap_after_harmony", pt.size = 0.8)
  p <- cowplot::plot_grid(plotlist =  list(p1,p2,p3,p4), nrow = 2, ncol = 2)
  ggsave(p, filename = paste(dir, "/umaps_markers/", gene, "_treated_control.pdf", sep = ""), dpi = 300, width = 32, height = 32)
})
```

# For lecture
```{r}
library(ggplot2)
temp <- data.object.filter.norm.integrated[, data.object.filter.norm.integrated$sample_id %in% "V11D06-099_A1"]
df <- data.frame(spots = c(1:length(colnames(temp))) , umi_count = temp$nCount_Spatial)
rownames(df) <- df$spots

median_c <- median(df$umi_count)

df$umi_count_scaled <- df$umi_count
for(i in 1:length(rownames(df))){
  scalefactor <- median_c / df$umi_count[i]
  df$umi_count_scaled[i] <- df$umi_count_scaled[i] * scalefactor
}

ggplot(data=df, aes(x=spots, y=umi_count)) +
  geom_line() + theme_classic()

ggplot(data=df, aes(x=spots, y=umi_count_scaled)) +
  geom_line() + theme_classic()
```
